backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
cattrs==23.2.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flatbuffers==24.3.25
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
numpy==2.0.0
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.11
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-cache==1.2.1
requests-mock==1.12.1
retry-requests==2.0.0
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.2
